+++ .gitignore
@@ -25,3 +25,4 @@ dist-ssr
 /test-results/
 /playwright-report/
 /playwright/.cache/
+coverage
